refac: Break up help.rs into smaller modules
