+++ src/help/mod.rs
@@ -1,0 +1,5 @@
+mod data;
+mod styles;
+mod term;
+
+pub use term::{print_auth_help, print_help, print_self_help};
